EN: User Management update with task CRUD (#2)
* Added User Management and user sesison management along with tasks, added middleware support

* Updated API contract documentation for todo and user API

* Enhanced server creation and configuration using .env file

+++ internal/store/store.go
@@ -8,6 +8,7 @@ import (
 	"time"
 	"todoapp/internal/models"
 
+	"github.com/google/uuid"
 	_ "modernc.org/sqlite"
 )
 
@@ -16,42 +17,150 @@ type Store struct {
 	Log *slog.Logger
 }
 
-func New(logger *slog.Logger) (*Store, error) {
+func New(logger *slog.Logger) (*Store, func(), error) {
 	const dbFile string = "./tasks.db"
 
 	db, err := sql.Open("sqlite", dbFile)
 	if err != nil {
 		logger.Error("unable to connect sqlite", "error", err.Error())
-		return nil, err
+		return nil, nil, err
 	}
 
-	const create string = `
-		CREATE TABLE IF NOT EXISTS tasks(
-		task_id TEXT NOT NULL PRIMARY KEY,
-		task_title TEXT NOT NULL,
-		done_status BOOLEAN NOT NULL CHECK (done_status IN (0, 1)),
-		added_at DATETIME NOT NULL,
-		modified_at DATETIME);`
-
-	if _, err := db.Exec(create); err != nil {
-		return nil, err
+	if err = db.Ping(); err != nil {
+		logger.Error("database not reachable", "error", err)
+		return nil, nil, err
 	}
 
-	return &Store{
-		Log: logger,
-		DB:  db,
-	}, nil
+	fn := func() {
+		if err := db.Close(); err != nil {
+			logger.Error("DB Close", "error", err)
+		}
+	}
+
+	if err := runMigration(db); err != nil {
+		return nil, fn, err
+	}
+
+	return &Store{Log: logger, DB: db}, fn, nil
 }
 
-func (s Store) GetAll(ctx context.Context) ([]models.Task, error) {
-	var res = make([]models.Task, 0)
+func (s *Store) RegisterUser(ctx context.Context, data *models.UserData, session *models.UserSession) (*models.UserSession, error) {
+	var (
+		err  error
+		opts = sql.TxOptions{Isolation: sql.LevelReadCommitted}
+	)
 
-	rows, err := s.DB.QueryContext(ctx, getAll)
+	tx, err := s.DB.BeginTx(ctx, &opts)
 	if err != nil {
 		return nil, err
 	}
 
-	s.Log.Debug("completed query with no error", "sql-query", getAll)
+	_, err = tx.ExecContext(ctx, registerQuery, data.ID, data.Name, data.Email, data.Password)
+	if err != nil {
+		return nil, rollback(tx, err)
+	}
+
+	_, err = tx.ExecContext(ctx, createSession, session.ID, session.UserID, session.Token, session.Expiry)
+	if err != nil {
+		return nil, rollback(tx, err)
+	}
+
+	if err := tx.Commit(); err != nil {
+		return nil, rollback(tx, err)
+	}
+
+	return session, nil
+}
+
+func (s *Store) CreateSession(ctx context.Context, session *models.UserSession) error {
+	_, err := s.DB.ExecContext(ctx, createSession, session.ID, session.UserID, session.Token, session.Expiry)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (s *Store) GetSessionByID(ctx context.Context, userID *uuid.UUID) (*models.UserSession, error) {
+	var session models.UserSession
+
+	if userID == nil {
+		return nil, errors.New("invalid user_id provided")
+	}
+
+	row := s.DB.QueryRowContext(ctx, getSession, *userID)
+	if err := row.Scan(&session.ID, &session.UserID, &session.Token, &session.Expiry); err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			return nil, models.ErrNotFound
+		}
+
+		return nil, err
+	}
+
+	return &session, nil
+}
+
+func (s *Store) RefreshSession(ctx context.Context, newSession *models.UserSession) (*models.UserSession, error) {
+	_, err := s.DB.ExecContext(ctx, refreshSession, newSession.Token, newSession.Expiry, newSession.ID)
+	if err != nil {
+		return nil, err
+	}
+
+	s.Log.Info("session is refreshed", "userID", newSession.UserID)
+
+	return newSession, nil
+}
+
+func (s *Store) GetByEmail(ctx context.Context, email string) (*models.UserData, error) {
+	var user models.UserData
+
+	row := s.DB.QueryRowContext(ctx, getUser, email)
+
+	if err := row.Scan(&user.ID, &user.Name, &user.Email, &user.Password); err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			return nil, models.ErrUserNotFound
+		}
+
+		return nil, err
+	}
+
+	return &user, nil
+}
+
+func (s *Store) Logout(ctx context.Context, token *uuid.UUID) error {
+	tx, err := s.DB.BeginTx(ctx, &sql.TxOptions{Isolation: sql.LevelReadCommitted})
+	if err != nil {
+		return err
+	}
+
+	var session models.UserSession
+
+	row := tx.QueryRowContext(ctx, "SELECT * FROM sessions where token=?", *token)
+	if err = row.Scan(&session.ID, &session.UserID, &session.Token, &session.Expiry); err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			return rollback(tx, models.ErrNotFound)
+		}
+	}
+
+	if _, err = tx.ExecContext(ctx, "DELETE FROM sessions WHERE id=?", session.ID); err != nil {
+		return rollback(tx, err)
+	}
+
+	if err := tx.Commit(); err != nil {
+		s.Log.Error(err.Error(), "incident", "logout commit")
+		return err
+	}
+
+	return nil
+}
+
+func (s *Store) GetAll(ctx context.Context, userID *uuid.UUID) ([]models.Task, error) {
+	var res = make([]models.Task, 0)
+
+	rows, err := s.DB.QueryContext(ctx, getAll, *userID)
+	if err != nil {
+		return nil, err
+	}
 
 	defer rows.Close()
 
@@ -61,7 +170,7 @@ func (s Store) GetAll(ctx context.Context) ([]models.Task, error) {
 			done int
 		)
 
-		err := rows.Scan(&task.ID, &task.Title, &done, &task.AddedAt, &task.ModifiedAt)
+		err := rows.Scan(&task.ID, &task.UserID, &task.Title, &done, &task.AddedAt, &task.ModifiedAt)
 		if err != nil {
 			return nil, err
 		}
@@ -79,10 +188,10 @@ func (s Store) GetAll(ctx context.Context) ([]models.Task, error) {
 	return res, nil
 }
 
-func (s Store) Create(ctx context.Context, id, title string) (*models.Task, error) {
-	addTS := time.Now()
+func (s *Store) Create(ctx context.Context, id, title string, userID *uuid.UUID) (*models.Task, error) {
+	addTS := time.Now().UTC()
 
-	query, values := genInsertQuery(id, title, addTS)
+	query, values := genInsertQuery(id, title, *userID, addTS)
 
 	s.Log.Debug("generated query", "sql-query", query, "values", values)
 
@@ -101,10 +210,10 @@ func (s Store) Create(ctx context.Context, id, title string) (*models.Task, erro
 	return &task, nil
 }
 
-func (s Store) Update(ctx context.Context, id, title string) (*models.Task, error) {
+func (s *Store) Update(ctx context.Context, id, title string, userID *uuid.UUID) (*models.Task, error) {
 	modifiedTS := time.Now()
 
-	query, values := genUpdateQuery(id, title, modifiedTS)
+	query, values := genUpdateQuery(id, title, *userID, modifiedTS)
 
 	s.Log.Debug("generated query", "sql-query", query, "values", values)
 
@@ -115,6 +224,7 @@ func (s Store) Update(ctx context.Context, id, title string) (*models.Task, erro
 
 	t := models.Task{
 		ID:         id,
+		UserID:     *userID,
 		Title:      title,
 		IsDone:     false,
 		ModifiedAt: &modifiedTS,
@@ -123,8 +233,8 @@ func (s Store) Update(ctx context.Context, id, title string) (*models.Task, erro
 	return &t, nil
 }
 
-func (s Store) Delete(ctx context.Context, id string) error {
-	_, err := s.DB.ExecContext(ctx, "DELETE FROM tasks WHERE task_id=?", id)
+func (s *Store) Delete(ctx context.Context, id string, userID *uuid.UUID) error {
+	_, err := s.DB.ExecContext(ctx, "DELETE FROM tasks WHERE task_id=? AND user_id=?", id, *userID)
 	if err != nil {
 		return err
 	}
@@ -132,27 +242,19 @@ func (s Store) Delete(ctx context.Context, id string) error {
 	return nil
 }
 
-func (s Store) MarkDone(ctx context.Context, id string) (*models.Task, error) {
+func (s *Store) MarkDone(ctx context.Context, id string, userID *uuid.UUID) (*models.Task, error) {
 	var (
 		task = models.Task{ID: id}
 		done int
 	)
 
-	res, err := s.DB.ExecContext(ctx, "UPDATE tasks SET done_status=? WHERE task_id=?", 1, id)
+	_, err := s.DB.ExecContext(ctx, "UPDATE tasks SET done_status=? WHERE task_id=? AND user_id=?", 1, id, *userID)
 	if err != nil {
 		return nil, err
 	}
 
-	val, err := res.RowsAffected()
-	if err != nil {
-		return nil, err
-	}
-
-	if val > 1 {
-		return nil, errors.New("not able to update")
-	}
-
-	row := s.DB.QueryRowContext(ctx, "SELECT task_title, done_status, added_at, modified_at FROM tasks WHERE task_id=?", id)
+	row := s.DB.QueryRowContext(ctx, `SELECT task_title, done_status, added_at, modified_at FROM tasks 
+		WHERE task_id=? AND user_id=?`, id, *userID)
 	if err := row.Scan(&task.Title, &done, &task.AddedAt, &task.ModifiedAt); err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
 			return nil, models.ErrNotFound
